refactor: move and rename routes
/count -> /get_activity_count
/workers -> /get_workers_status
