+++ internal/node/component/info/handler_workers_test.go
@@ -1,4 +1,4 @@
-package decentralized
+package info
 
 import (
 	"testing"
