+++ internal/node/node.go
@@ -52,7 +52,7 @@ func NewCoreService(ctx context.Context, config *config.File, databaseClient dat
 
 	apiServer.Use(middleware.CORSWithConfig(middleware.DefaultCORSConfig))
 
-	infoComponent := info.NewComponent(ctx, apiServer, config)
+	infoComponent := info.NewComponent(ctx, apiServer, config, databaseClient, redisClient)
 	node.components = append(node.components, &infoComponent)
 
 	if len(config.Component.RSS) > 0 {
